refactor code; SearchBar moved into FilterableProductTable
