+++ src/App.js
@@ -10,34 +10,17 @@ import './App.css';
 
 // importing components
 import FilterableProductTable from './components/organisms/FilterableProductTable';
-import SearchBar from './components/molecules/SearchBar';
 
-class App extends React.Component {
+function App() {
 
-  state = {
-    dataArr: productDataObject.data,
-    searchTerm: ""
+  return (
+    <div className="App container">
+      <h1 className="h1">IronStore</h1>
+      <FilterableProductTable products={productDataObject.data}></FilterableProductTable>
+
+    </div>
+  );
 }
 
-searchHandler = (searchTerm) => {
-    this.setState({
-        searchTerm: searchTerm
-    })
-}
-  render() {
-
-           let filteredProducts = this.state.dataArr.filter((product) => product.name.toLowerCase().includes(this.state.searchTerm))
-
-    return (
-      <div className="App container">
-        <h1 className="h1">IronStore</h1>
-        <SearchBar onSearchCallback={this.searchHandler} currentSearchTerm={this.state.searchTerm}></SearchBar>
-
-        <FilterableProductTable products={filteredProducts}></FilterableProductTable>
-        
-      </div>
-    );
-  }
-}
 
 export default App;
